DX-1000 implement class comment

+++ src/main/java/de/microtema/maven/plugin/contract/java/template/JavaTemplate.java
@@ -66,7 +66,7 @@ public class JavaTemplate {
             case TEXT:
                 return "String";
             case NUMBER:
-                return "Integer";
+                return "Long";
             case BOOLEAN:
                 return "Boolean";
             default:
